Move Steam-related types under Steam namespace.
