+++ toofz.NecroDancer.Leaderboards.Tests/Steam/UgcHttpClientTests.cs
@@ -4,10 +4,11 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RichardSzalay.MockHttp;
+using toofz.NecroDancer.Leaderboards.Steam;
 using toofz.NecroDancer.Leaderboards.Tests.Properties;
 using toofz.TestsShared;
 
-namespace toofz.NecroDancer.Leaderboards.Tests
+namespace toofz.NecroDancer.Leaderboards.Tests.Steam
 {
     class UgcHttpClientTests
     {
