+++ toofz.NecroDancer.Leaderboards/Steam/IUgcHttpClient.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace toofz.NecroDancer.Leaderboards
+namespace toofz.NecroDancer.Leaderboards.Steam
 {
     public interface IUgcHttpClient
     {
