+++ toofz.NecroDancer.Leaderboards/Steam/UgcHttpClient.cs
@@ -7,7 +7,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Threading.Tasks.Dataflow;
 
-namespace toofz.NecroDancer.Leaderboards
+namespace toofz.NecroDancer.Leaderboards.Steam
 {
     public sealed class UgcHttpClient : HttpClient, IUgcHttpClient
     {
